+++ src/Configuration/YoutubeTwigConfig.php
@@ -1,0 +1,26 @@
+<?php
+
+/*
+ * Copyright (c) 2018 Heimrich & Hannot GmbH
+ * @license LGPL-3.0-or-later
+ */
+
+namespace HeimrichHannot\YoutubeBundle\Configuration;
+
+class YoutubeTwigConfig extends YoutubeConfig
+{
+    /**
+     * @var string
+     */
+    protected $template = '@HeimrichHannotContaoYoutube/youtube_video/youtube_video_default.html.twig';
+
+    /**
+     * @var string
+     */
+    protected $privacyTemplate = '@HeimrichHannotContaoYoutube/youtube_privacy/youtube_privacy_default.html.twig';
+
+    /**
+     * @var string
+     */
+    protected $modalTemplate = '@HeimrichHannotContaoYoutube/youtube_modalvideo/youtube_modalvideo_default.html.twig';
+}