+++ src/Resources/contao/languages/de/tl_reader_config_element.php
@@ -1,0 +1,10 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_reader_config_element'];
+
+/**
+ * Reference
+ */
+$lang['reference'] = [
+    \HeimrichHannot\YoutubeBundle\ConfigElementType\YoutubeReaderConfigElementType::TYPE => "YouTube",
+];