+++ src/Resources/contao/languages/en/tl_settings.php
@@ -1,0 +1,11 @@
+<?php
+
+/**
+ * Fields
+ */
+$GLOBALS['TL_LANG']['tl_settings']['youtubeApiKey']           = ['YouTube-Api-Key', 'Enter the API key here.'];
+$GLOBALS['TL_LANG']['tl_settings']['youtubeSkipImageCaching'] = ['Don\'t cache Youtube preview images', 'Don\'t cache preview images for Youtube videos.'];
+/**
+ * Legends
+ */
+$GLOBALS['TL_LANG']['tl_settings']['youtube_legend'] = 'YouTube';